# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiChart API",
    "summary": "Backend API for the NiChart medical-imaging pipeline platform.",
    "description": "Provides endpoints for managing projects, uploading imaging data, running containerised processing pipelines (locally via Docker or on AWS Batch), and retrieving results. \n\n**Authentication**: In cloud mode every request requires a Cognito ID token in the ``Authorization: Bearer`` header. In local mode (``NICHART_EXECUTION_MODE=local``) authentication is bypassed.",
    "contact": {
      "name": "CBICA",
      "url": "https://github.com/CBICA/NiChart_Project"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Begin OAuth2 sign-in",
        "description": "Generates a CSRF ``state`` and PKCE ``code_verifier``/``code_challenge``, stores them in short-lived httpOnly cookies, then redirects the browser to the Cognito Hosted UI authorization endpoint. Navigate the browser (not a fetch/XHR) to this URL to start sign-in.",
        "operationId": "login_auth_login_get",
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "302": {
            "description": "Redirect to Cognito authorize endpoint."
          },
          "503": {
            "description": "Cognito domain not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/auth/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "OAuth2 callback \u2014 exchange code for session",
        "description": "Cognito redirects here after the user authenticates. Validates the CSRF ``state`` cookie, exchanges the authorization code for tokens (using the PKCE verifier + client secret), verifies the ID token, sets the ``session`` cookie, and redirects to the frontend. This endpoint is called by the browser following a redirect \u2014 not by JS fetch.",
        "operationId": "callback_auth_callback_get",
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "302": {
            "description": "Redirect to frontend after successful authentication."
          },
          "400": {
            "description": "State mismatch or missing parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "401": {
            "description": "Token validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "502": {
            "description": "Cognito token exchange failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Return current user claims",
        "description": "Reads the ``session`` cookie and returns safe user claims if the session is valid. The frontend should poll this endpoint to determine auth state on page load. Returns 401 when there is no session or it has expired \u2014 the client should redirect to ``GET /auth/login``. In local mode, always returns the OS user without a cookie.",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          },
          "401": {
            "description": "No valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign out",
        "description": "Clears the ``session`` and ``refresh_token`` cookies, then redirects to the Cognito logout URL (which itself redirects to the configured frontend URL). Navigate the browser to this URL \u2014 do not call it via fetch/XHR.",
        "operationId": "logout_auth_logout_get",
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "302": {
            "description": "Redirect to Cognito logout endpoint."
          }
        }
      }
    },
...
    "/catalog/docs/{docs_id}": {
      "get": {
        "tags": [
          "Docs"
        ],
        "summary": "Get documentation topic manifest",
        "description": "Returns the manifest for a documentation topic: its sections, audience tags, and the pipeline IDs it covers. Use the ``file`` field of each section with ``GET /catalog/docs/{docs_id}/{file}`` to fetch the actual content. No authentication required.",
        "operationId": "get_docs_manifest_catalog_docs__docs_id__get",
        "parameters": [
          {
            "name": "docs_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Docs Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocManifest"
                }
              }
            }
          },
          "404": {
            "description": "Topic or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/catalog/docs/{docs_id}/{path}": {
      "get": {
        "tags": [
          "Docs"
        ],
        "summary": "Fetch a documentation file",
        "description": "Serves any file within the documentation topic folder: markdown prose, images, or JSON data. Relative image references in markdown files resolve naturally to further requests under this same path prefix. Path traversal outside the topic folder is rejected. No authentication required.",
        "operationId": "get_docs_file_catalog_docs__docs_id___path__get",
        "parameters": [
          {
            "name": "docs_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Docs Id"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "Topic or file not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocManifest": {
        "properties": {
          "docs_id": {
            "type": "string",
            "title": "Docs Id",
            "description": "Topic identifier (folder name under resources/docs/)."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Human-readable topic title."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "One-sentence summary shown in the docs index."
          },
          "pipelines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pipelines",
            "description": "Pipeline IDs covered by this documentation topic. Multiple pipelines (e.g. harmonized and base variants) can share one topic."
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail",
            "description": "Filename of a thumbnail image relative to the topic folder. Fetch with GET /catalog/docs/{docs_id}/{thumbnail}."
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/DocSection"
            },
            "type": "array",
            "title": "Sections",
            "description": "Ordered list of content sections available for this topic."
          }
        },
        "type": "object",
        "required": [
          "docs_id",
          "title"
        ],
        "title": "DocManifest",
        "description": "Full documentation manifest for a single topic."
      },
      "DocSection": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for this section within the topic."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Display title for this section (e.g. tab label)."
          },
          "file": {
            "type": "string",
            "title": "File",
            "description": "Filename relative to the topic folder. Fetch the content with GET /catalog/docs/{docs_id}/{file}."
          },
          "audience": {
            "type": "string",
            "enum": [
              "user",
              "developer",
              "all"
            ],
            "title": "Audience",
            "description": "'user' \u2014 shown in the NiChart end-user interface. 'developer' \u2014 shown in developer/technical documentation. 'all' \u2014 shown in both contexts.",
            "default": "all"
          },
          "type": {
            "type": "string",
            "enum": [
              "markdown",
              "data",
              "image"
            ],
            "title": "Type",
            "description": "'markdown' \u2014 render as prose with react-markdown or equivalent. 'data' \u2014 parse as JSON for plot/chart rendering. 'image' \u2014 display as a static image.",
            "default": "markdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "file"
        ],
        "title": "DocSection",
        "description": "A single section within a documentation topic."
      },
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable description of the error."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Standard error response body."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserInfo": {
        "properties": {
          "sub": {
            "type": "string",
            "title": "Sub"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Groups",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "sub"
        ],
        "title": "UserInfo",
        "description": "Safe user claims returned by GET /auth/me. The raw token is never included."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}